Location Name Updates
Updates the naming convention for locations. This makes locations easier to read as well as fixes some overflow issues within the Exclude Locations menu due to the length of some location names, it will help in the future with the in-game spoiler log for the same reason.
Some locations have also been slightly renamed for clarity.